requirements.txt
attrs==25.4.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
Flask==2.3.3
flask-cors==6.0.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
PyJWT==2.10.1

PyMySQL==1.1.2
python-dotenv==1.0.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
Werkzeug==3.1.3
zipp==3.23.0
